+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * `produceAtomic` coroutine builder. An alternative to `produce` which guarantee that the block is always invoked.
 * `parallel` operator to easily build parallel pipelines.
 * `parallelMap` and `parallelFilter` operators.
+* `produce(Iterable)` and `produce(Sequence)` functions. They replace `Iterable.openSubscription`, `Sequence.openSubscription` which are deprecated.
  
 ### Dependencies
 * Update Kotlin from `1.3.0-rc-57` to `1.3.10`